requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.16.1
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
email-validator==2.1.0.post1
filelock==3.18.0
Flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
gunicorn==21.2.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.6
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
numpy==1.24.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.8.1.78
packaging==25.0
pandas==2.3.0
phonenumbers==8.13.26
Pillow==10.0.1
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0num
PyJWT==2.8.0
pyparsing==3.0.9
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.0.4
requests==2.31.0
scipy==1.15.3
seaborn==0.12.2
six==1.17.0
SQLAlchemy==2.0.41
sympy==1.14.0
thop==0.1.1.post2209072238
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
twilio==8.10.0
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.155
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==2.3.7
yarl==1.20.1
